(134 & 145) Corr/incorr feedback (3-tries) & Data record tracking placeholders
The three incorrect answers-logic is in place, providing different feedback based on how many inccorect answers the users provide, moves on to the next clothing item after 3 incorr. tries. No 50/50 for item to be incorrect, now ClothingID is checked at the PelleContainer to see if it matches the prompt.

Some data tracking based on the players performance is stored eagerly in levelOneView (can be moved if we start implementing level2 and level3.